move align codes to algo.__init__
